타입 수정 숫자 범위가 문제였음..

+++ src/main/java/com/ki/level1/p12954/Main.java
@@ -7,15 +7,15 @@ public class Main {
 }
 
 class Solution {
-    public int[] solution(int x, int n) {
+    public long[] solution(int x, int n) {
         if (n == 0) {
-            return new int[0];
+            return new long[0];
         }
 
-        int[] answer = new int[n];
+        long[] answer = new long[n];
 
         for (int i = 0; i < n; i++) {
-            answer[i] = x * (i + 1);
+            answer[i] =(long) x * (i + 1);
         }
 
         return answer;
